Cleaned up docs examples to be just one file
