+++ README.md
@@ -47,7 +47,7 @@ it("displays the title when there is a title", () => {
 });
 ```
 
-See [Selectors](./docs/Selectors.md) for more documentation or [Heading](./docs/examples/Heading/Heading.test.tsx) for a code example.
+See [Selectors](./docs/Selectors.md) for more documentation or [Heading](./docs/examples/Heading.test.tsx) for a code example.
 
 ### Mocking Selectors
 
@@ -73,7 +73,7 @@ it("displays the title when there is a title", () => {
 
 If you'd like more control over the return values, you can use `.giveMock` to provide a [Jest mock](https://jestjs.io/docs/en/mock-functions.html).
 
-See [Selectors](./docs/Selectors.md) for more documentation or [Heading](./docs/examples/Heading/Heading.test.tsx) for a code example.
+See [Selectors](./docs/Selectors.md) for more documentation or [Heading](./docs/examples/Heading.test.tsx) for a code example.
 
 ### Dispatch Spies
 
@@ -95,7 +95,7 @@ it("dispatches the pageLoaded action when rendered", () => {
 });
 ```
 
-See [Dispatches](./docs/Dispatches.md) for more documentation or [Clicker](./docs/examples/Clicker/Clicker.test.tsx) for a code example.
+See [Dispatches](./docs/Dispatches.md) for more documentation or [Clicker](./docs/examples/Clicker.test.tsx) for a code example.
 
 ## Gotchas
