--- docs/examples/Clicker/redux.ts
@@ -1,3 +1,0 @@
-export const MARK_CLICKED = "MARK_CLICKED";
-
-export const markClicked = () => ({ type: MARK_CLICKED });
