--- docs/examples/Heading/redux.ts
@@ -1,5 +1,0 @@
-export type RootState = {
-  title: string;
-};
-
-export const selectTitle = (state: RootState) => state.title;
